media: dt-bindings: cedrus: Make allwinner,sram optional
Allwinner SoCs can remap some bits of peripheral SRAM to a contiguous
range of addresses for use by early boot software. Usually the video
engine's SRAM is used for this purpose, so its mapping must be switched
back before the hardware can be used.

However, the D1 and related SoCs use the DSP SRAM for this purpose. As a
result, the video engine SRAM is not switchable, and there is no need
for an allwinner,sram reference in the devicetree.

Signed-off-by: Samuel Holland <samuel@sholland.org>