irqchip/sifive-plic: Enable oneshot-safe threaded EOIs
This defers the EOI until the IRQ thread has finished, avoiding the need
to mask the IRQ while the thread is pending.

Cover-letter:
genirq: oneshot-safe threaded EOIs
This is the optimization that I promised back in
https://lore.kernel.org/lkml/20220701202440.59059-1-samuel@sholland.org/
and got back around to cleaning up. To quote that cover letter:

"A further optimization is to take advantage of the fact that multiple
IRQs can be claimed at once. This allows removing the mask operations
for oneshot IRQs -- i.e. the combination of IRQCHIP_ONESHOT_SAFE and
IRQCHIP_EOI_THREADED, which is not currently supported."

Or in other words, we reuse the oneshot infrastructure for EOIs instead
of masking.

There are a few functions/variables that probably should be renamed now
that I have muddied what "oneshot" means. I elected not to do that in
this version to avoid distracting from the functional changes.
END

Series-to: Marc Zyngier <maz@kernel.org>
Series-to: Thomas Gleixner <tglx@linutronix.de>
Series-to: Palmer Dabbelt <palmer@dabbelt.com>
Series-cc: linux-riscv@lists.infradead.org

Signed-off-by: Samuel Holland <samuel@sholland.org>