riscv: dts: allwinner: Hook up system LDOs
All supported boards use LDOA for the board's 1.8 V rail. D1s boards use
LDOB to power the in-package DRAM; this LDO is unused on D1 boards.

Commit-changes: 6
 - New patch for v6, now that the rest of the DT is upstream

Cover-changes: 6
 - Drop the SRAM/system control driver patch, which was merged

Series-version: 6

Cover-letter:
regulator: Add support for Allwinner D1 system LDOs
This series adds the binding and driver for one of the two pairs of LDOs
inside the Allwinner D1 SoC family. I am splitting up the two pairs of
LDOs into separate series; the analog LDOs depend on the (unfinished)
audio codec binding, but they are not required to boot any board.

A binding change is required for the SRAM controller to accept the
regulators device as its child. The example for the regulator device
binding is in SRAM controller binding document, per Rob's request to
keep MFD examples in one place.
END

Series-to: Liam Girdwood <lgirdwood@gmail.com>
Series-to: Mark Brown <broonie@kernel.org>

Signed-off-by: Samuel Holland <samuel@sholland.org>